Improve and simplyfy plugin Move Subtitles
Up to now, new start was like the current position. Now, when video is loaded,
new position is prefilled based on current video player position.

Also inform user when nothing is moved because new and old start are the
same and improve files comments and the description of the action on
menu hover.

Up until now, there was a user configuration whether only selected
subtitles are to be moved, or the first selected and all subsequent
subtitles are to be moved. This is getting rid of this setting and
instead decides based on selection. When one subtitle is selected, this
subtitled and all subsequent subtitles are moved. When more subtitles
are selected, only those are moved.