Join Document Plugin: Add Append action and make offset configurable
This adds new action that adds subtitles from another
file without changing timecodes at all. This is useful if the subtitles
are for the same video, for example if two people were working on the
same video and now need to join their separate parts.

Also, the offset of the original Join subtitle action is not
configurable via a dialog. The dialog is also smart, as it offers
the duration of video as the default offset if a video is open. When no
video is open, the endtime of last subtitle is offered, mimicking
current behaviour. Of course, the user can also choose something else.

In practice, when joining subtitles like this, often they will be timed
to a different video, that is supposedly going to be joined after the
current video.

Also a gramatical fix to translation files is applied.